Create vam.py
vam.py generates a binary matrix from a table of variants and a .bed file of annotations, with rows representing variants and columns representing annotations. Entries of the matrix are 1 if the annotation overlaps the variant, 0 if it does not. Generating FGWAS input files is one use case for this tool.